refactor(delivery): give routing a module of its own
Where a batch goes belongs to the batch, not to whichever window asked -- the
review view already said so in a comment on the function it had to expose when a
second surface needed the same answer. Holding it on the largest module in the
plugin made the composer load a diff view to read a label.

The target, the wording for it, and the picker dispatch move to a module named
for what they do. The split with the review view is routing against windows: the
choice and the adapter call are delivery's, the origin window, the fallback into
the diff when it dies mid-pick, and the winbar repaint stay with the surface that
owns them.

The payload handoff and submit stay where they are; they read the target from its
new home and move in their own slices.
